requirements/base.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.3.0
django-js-asset==1.1.0
django-mptt==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.1
